+++ packages/transaction-controller/src/utils/batch.test.ts
@@ -472,14 +472,6 @@ describe('Batch Utils', () => {
       );
     });
 
-    it('throws if chain not supported', async () => {
-      doesChainSupportEIP7702Mock.mockReturnValueOnce(false);
-
-      await expect(addTransactionBatch(request)).rejects.toThrow(
-        rpcErrors.internal('Chain does not support EIP-7702'),
-      );
-    });
-
     it('throws if no public key', async () => {
       doesChainSupportEIP7702Mock.mockReturnValueOnce(true);
 
@@ -662,6 +654,7 @@ describe('Batch Utils', () => {
         mockRequestApproval(MESSENGER_MOCK, {
           state: 'approved',
         });
+        doesChainSupportEIP7702Mock.mockReturnValueOnce(false);
       });
 
       it('adds each nested transaction', async () => {
@@ -675,7 +668,7 @@ describe('Batch Utils', () => {
         addTransactionBatch({
           ...request,
           publishBatchHook,
-          request: { ...request.request, useHook: true },
+          request: { ...request.request },
         }).catch(() => {
           // Intentionally empty
         });
@@ -732,7 +725,7 @@ describe('Batch Utils', () => {
           addTransactionBatch({
             ...request,
             publishBatchHook,
-            request: { ...request.request, useHook: true, origin },
+            request: { ...request.request, origin },
           }).catch(() => {
             // Intentionally empty
           });
@@ -786,7 +779,7 @@ describe('Batch Utils', () => {
         addTransactionBatch({
           ...request,
           publishBatchHook,
-          request: { ...request.request, useHook: true },
+          request: { ...request.request },
         }).catch(() => {
           // Intentionally empty
         });
@@ -865,7 +858,7 @@ describe('Batch Utils', () => {
         addTransactionBatch({
           ...request,
           publishBatchHook,
-          request: { ...request.request, useHook: true },
+          request: { ...request.request },
         }).catch(() => {
           // Intentionally empty
         });
@@ -948,7 +941,6 @@ describe('Batch Utils', () => {
               },
               request.request.transactions[1],
             ],
-            useHook: true,
           },
         }).catch(() => {
           // Intentionally empty
@@ -1047,7 +1039,6 @@ describe('Batch Utils', () => {
               },
               request.request.transactions[1],
             ],
-            useHook: true,
           },
         }).catch(() => {
           // Intentionally empty
@@ -1098,7 +1089,7 @@ describe('Batch Utils', () => {
         const resultPromise = addTransactionBatch({
           ...request,
           publishBatchHook,
-          request: { ...request.request, useHook: true },
+          request: { ...request.request },
         });
 
         resultPromise.catch(() => {
@@ -1160,7 +1151,6 @@ describe('Batch Utils', () => {
           publishBatchHook,
           request: {
             ...request.request,
-            useHook: true,
             requireApproval: false,
           },
         }).catch(() => {
@@ -1217,7 +1207,6 @@ describe('Batch Utils', () => {
           publishBatchHook,
           request: {
             ...request.request,
-            useHook: true,
             requireApproval: false,
           },
         }).catch(() => {
@@ -1244,6 +1233,10 @@ describe('Batch Utils', () => {
     });
 
     describe('with sequential publish batch hook', () => {
+      beforeEach(() => {
+        doesChainSupportEIP7702Mock.mockReturnValueOnce(false);
+      });
+
       let sequentialPublishBatchHook: jest.MockedFn<PublishBatchHook>;
 
       beforeEach(() => {
@@ -1332,7 +1325,6 @@ describe('Batch Utils', () => {
           publishBatchHook: undefined,
           request: {
             ...request.request,
-            useHook: true,
             requireApproval: false,
           },
         }).catch(() => {
@@ -1357,7 +1349,7 @@ describe('Batch Utils', () => {
           addTransactionBatch({
             ...request,
             publishBatchHook: undefined,
-            request: { ...request.request, useHook: true },
+            request: { ...request.request },
           }),
         ).rejects.toThrow('Test error');
 
@@ -1375,7 +1367,7 @@ describe('Batch Utils', () => {
           ...request,
           publishBatchHook: undefined,
           messenger: MESSENGER_MOCK,
-          request: { ...request.request, useHook: true, origin: ORIGIN_MOCK },
+          request: { ...request.request, origin: ORIGIN_MOCK },
         }).catch(() => {
           // Intentionally empty
         });
@@ -1415,7 +1407,6 @@ describe('Batch Utils', () => {
           messenger: MESSENGER_MOCK,
           request: {
             ...request.request,
-            useHook: true,
             origin: ORIGIN_MOCK,
           },
         }).catch(() => {
@@ -1490,6 +1481,10 @@ describe('Batch Utils', () => {
   });
 
   describe('isAtomicBatchSupported', () => {
+    beforeEach(() => {
+      jest.resetAllMocks();
+    });
+
     it('includes all feature flag chains if chain IDs not specified', async () => {
       getEIP7702SupportedChainsMock.mockReturnValueOnce([
         CHAIN_ID_MOCK,