+++ packages/transaction-controller/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `transactionBatches` array to state.
   - Add `TransactionBatchMeta` type.
 
+### Changed
+
+- Default addTransactionBatch to EIP7702 if supported, otherwise use sequential batch ([5853](https://github.com/MetaMask/core/pull/5853))
+
 ### Fixed
 
 - Support leading zeroes in `authorizationList` properties ([#5830](https://github.com/MetaMask/core/pull/5830))